Add first conversation with Jade Champion in Holy City before the Cocoon dungeon is completed.
