+++ app/content/dialogue/jadeChampion.ts
@@ -1,13 +1,11 @@
 import { burstAnimation, FieldAnimationEffect } from 'app/content/effects/animationEffect';
 import { dialogueHash } from 'app/content/dialogue/dialogueHash';
 import { FRAME_LENGTH } from 'app/gameConstants';
-import { appendCallback, appendScript, runPlayerBlockingCallback } from 'app/scriptEvents';
+import { appendCallback, appendScript, runBlockingCallback, runPlayerBlockingCallback } from 'app/scriptEvents';
 import { createObjectInstance } from 'app/utils/createObjectInstance';
 import { moveNPCToTargetLocation } from 'app/utils/npc';
 import { addObjectToArea, removeObjectFromArea } from 'app/utils/objects';
 import { addEffectToArea } from 'app/utils/effects';
-import { enterZoneByTarget } from 'app/utils/enterZoneByTarget';
-import {findObjectInstanceById} from 'app/utils/findObjectInstanceById';
 
 function addBurstEffect(this: void, state: GameState, npc: NPC, area: AreaInstance): void {
     const hitbox = npc.getHitbox();
@@ -21,368 +19,119 @@ function addBurstEffect(this: void, state: GameState, npc: NPC, area: AreaInstan
     addEffectToArea(state, area, animation);
 }
 
-function npcAndHeroTeleportAnimation(state: GameState, npc: NPC) {
-    appendScript(state, '{wait:500}');
-    runPlayerBlockingCallback(state, (state: GameState) => {
-        npc.animationTime += FRAME_LENGTH;
-        if (npc.currentAnimationKey !== 'idle') {
-            npc.changeToAnimation('idle');
-            return true;
-        }
-    });
-    appendScript(state, '{wait:100}');
-    runPlayerBlockingCallback(state, (state: GameState) => {
-        state.hero.animationTime += FRAME_LENGTH;
-        state.hero.d = 'down';
-        if (state.hero.action !== 'kneel') {
-            state.hero.action = 'kneel';
-            return true;
-        }
-    });
-    appendScript(state, '{wait:500}');
-    runPlayerBlockingCallback(state, (state: GameState) => {
-        npc.animationTime += FRAME_LENGTH;
-        if (npc.currentAnimationKey !== 'cast') {
-            npc.changeToAnimation('cast');
-            return true;
-        }
-    });
-    appendScript(state, '{wait:400}');
-    appendCallback(state, (state: GameState) => {
-        addBurstEffect(state, npc, state.hero.area);
-        state.hero.renderParent = npc;
-        removeObjectFromArea(state, npc);
-    });
-    appendScript(state, '{wait:700}');
-}
-
-dialogueHash.jadeChampionWarTemple = {
-    key: 'jadeChampionWarTemple',
+dialogueHash.jadeChampion = {
+    key: 'jadeChampion',
     mappedOptions: {
-        warTempleEncounter: (state: GameState) => {
-            appendScript(state, '{playTrack:village}{wait:100');
+        firstConversation: (state: GameState) => {
             // add Jade Champion to the screen
             const jadeChampion = createObjectInstance(state, {
                 id: 'jadeChampion',
-                dialogueKey: 'jadeChampionWarTemple',
+                dialogueKey: 'jadeChampion',
                 status: 'normal',
-                x: 86,
-                y: 474,
+                x: 200,
+                y: 216, // initially appear offscreen, walk south
                 type: 'npc',
                 behavior: 'none',
                 style: 'jadeChampion',
                 d: 'down',
             }) as NPC;
-            jadeChampion.speed = 1.75;
+            jadeChampion.speed = 1.5;
             appendCallback(state, (state: GameState) => {
                 addObjectToArea(state, state.hero.area, jadeChampion);
             });
+            // Move the player to a good x,y position before talking to the Jade Champion.
+            // target coordinates: { x: 134, y: 446 }
+            // FIX, TODO: camera movement is jerky and weird.
+            runBlockingCallback(state, (state: GameState) => {
+                const hero = state.hero;
+                // fix x coordinate
+                if (hero.x === 134) {
+                    hero.d = 'up';
+                    delete hero.action;
+                    return false;
+                }
+                hero.action = 'walking';
+                hero.animationTime += FRAME_LENGTH;
+                if (hero.x < 134) {
+                    hero.d = 'right';
+                    hero.x = Math.min(hero.x + 1, 134);
+                } else {
+                    hero.d = 'left';
+                    hero.x = Math.max(hero.x - 1, 134);
+                }
+                return true;
+            });
+            runBlockingCallback(state, (state: GameState) => {
+                const hero = state.hero;
+                // fix y coordinate
+                if (hero.y === 446) {
+                    hero.d = 'up';
+                    delete hero.action;
+                    return false;
+                }
+                hero.action = 'walking';
+                hero.animationTime += FRAME_LENGTH;
+                if (hero.y < 446) {
+                    hero.d = 'down';
+                    hero.y = Math.min(hero.y + 1, 446);
+                } else {
+                    hero.d = 'up';
+                    hero.y = Math.max(hero.y - 1, 446);
+                }
+                return true;
+            });
+            // JC walks onscreen from the north
+            // start: { x: 200, y: 216 }
+            // 1: { x: 200, y: 348 }
+            // 2: { x: 142, y: 348 }
+            // end: { x: 142, y: 424 }
+            // 1
             runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.speed = 1;
                 jadeChampion.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, jadeChampion, 162, 476, 'move')) {
+                if (moveNPCToTargetLocation(state, jadeChampion, 200, 348, 'move')) {
                     return true;
                 }
                 jadeChampion.changeToAnimation('idle');
             });
+            // 2
+            runPlayerBlockingCallback(state, (state: GameState) => {
+                jadeChampion.animationTime += FRAME_LENGTH;
+                if (moveNPCToTargetLocation(state, jadeChampion, 142, 348, 'move')) {
+                    return true;
+                }
+                jadeChampion.changeToAnimation('idle');
+            });
+            // end
+            runPlayerBlockingCallback(state, (state: GameState) => {
+                jadeChampion.animationTime += FRAME_LENGTH;
+                if (moveNPCToTargetLocation(state, jadeChampion, 142, 424, 'move')) {
+                    return true;
+                }
+                jadeChampion.d = 'down';
+                jadeChampion.changeToAnimation('idle');
+            });
+            appendScript(state, `You're that Vanara child the guards caught sneaking into the temple.
+                {|}You really should know better than to come here by now.
+                {|}It won't be my fault if you get chased out of town again.
+                {|}Just stay away from the Spirit Gate or I'll have to kick you out myself.`);
+            // JC teleports away
             appendScript(state, '{wait:500}');
-            appendScript(state, `Stop right there!`);
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.speed = 1.75;
-                jadeChampion.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, jadeChampion, 234, 476, 'move')) {
-                    return true;
-                }
-                jadeChampion.changeToAnimation('idle');
-            });
             runPlayerBlockingCallback(state, (state: GameState) => {
                 jadeChampion.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, jadeChampion, 234, 394, 'move')) {
+                if (jadeChampion.currentAnimationKey !== 'cast') {
+                    jadeChampion.changeToAnimation('cast');
                     return true;
                 }
-                jadeChampion.d = 'right';
-                jadeChampion.changeToAnimation('idle');
             });
-            appendScript(state, `
-                What did you do to the Spirit Beasts?
-                {|}Why would a Vanara child go poking around in the Spirit World?
-                {|}I'm sorry, but you'll have to explain yourself to the Spirit King.
-                {|}Come with me.
-                `
-            );
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, jadeChampion, 256, 394, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'up';
-                jadeChampion.changeToAnimation('idle');
-            });
-            npcAndHeroTeleportAnimation(state, jadeChampion);
-            // ****
-            // Grand Priest Conversation
-            // ****
-            let grandPriest: NPC;
+            appendScript(state, '{wait:400}');
             appendCallback(state, (state: GameState) => {
-                enterZoneByTarget(state, 'grandTemple', 'grandPriestMarker', null, false, (state) => {
-                    if (state.hero.renderParent && state.hero.renderParent.area !== state.hero.area) {
-                        delete state.hero.renderParent;
-                    }
-                    jadeChampion.x = 250;
-                    jadeChampion.y = 346;
-                    jadeChampion.speed = 1.25;
-                    addObjectToArea(state, state.hero.area, jadeChampion);
-                    grandPriest = findObjectInstanceById(state.hero.area, 'grandPriest') as NPC;
-                });
-            });
-            appendScript(state, '{wait:100}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, jadeChampion, 220, 346, 'move')) {
-                    return true;
-                }
-                jadeChampion.changeToAnimation('idle');
-            });
-            appendScript(state, '{wait:50}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, jadeChampion, 254, 300, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'up';
-                jadeChampion.changeToAnimation('idle');
-            });
-            appendScript(state, '{wait:200}');
-            appendScript(state, `I have brought you the Vanara child, Grand Priest.`);
-            appendCallback(state, (state: GameState) => {
-                state.hero.d = 'up';
-            });
-            appendScript(state, '{wait:500}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, jadeChampion, 208, 276, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'right';
-                jadeChampion.changeToAnimation('idle');
-            });
-            // UPDATE DIALOGUE
-            // grandPriest: What do you know of the disappearance of the Spirit Beasts, child?
-            // I can see that you don't know of what I speak.
-            // JC: But this child was in the War Temple when the Beasts vanished!
-            // That can't be a coincidence.
-            // priest to JC: Peace, Champion. Nothing that happens is merely a coincidence.
-            // That doesn't mean that this Vanara is responsible.
-            // priest to hero: Child, the Spirit King tells me to set you free.
-            // JC: How can this be? A Vanara child will be left to freely roam the Spirit World?
-            // priest: Silence! My commands are beyond the understanding of mortals.
-            // Champion, seek out the Spirit Beasts. The Spirit Gods shall watch the Vanara child themselves.
-            // Return this Vanara to the borders of our Holy City. This is my command.
-            // JC: Yes, Grand Priest.
-            // JC approaches the hero and teleports them to the Holy City
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                grandPriest.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, grandPriest, 256, 300, 'move')) {
-                    return true;
-                }
-                grandPriest.d = 'down';
-                grandPriest.changeToAnimation('idle');
-            });
-            appendScript(state, `What do you know of the disappearance of the Spirit Beasts, child?
-                {|}I can see that you don't know of what I speak.`);
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                grandPriest.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, grandPriest, 256, 250, 'move')) {
-                    return true;
-                }
-                grandPriest.d = 'down';
-                grandPriest.changeToAnimation('idle');
-            });
-            appendScript(state, '{wait:500}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                jadeChampion.speed = 1.25;
-                if (moveNPCToTargetLocation(state, jadeChampion, 256, 300, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'up';
-                jadeChampion.changeToAnimation('idle');
-            });
-            appendScript(state, `But this child was in the War Temple when the Beasts vanished!
-                {|}That can't be a coincidence.`);
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                grandPriest.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, grandPriest, 256, 270, 'move')) {
-                    return true;
-                }
-                grandPriest.d = 'down';
-                grandPriest.changeToAnimation('idle');
-            });
-            appendScript(state, `Peace, Champion. Nothing that happens is merely a coincidence.
-                {|}That doesn't mean that this Vanara is responsible.`);
-            appendScript(state, '{wait:500}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                jadeChampion.speed = 1;
-                if (moveNPCToTargetLocation(state, jadeChampion, 208, 276, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'right';
-                jadeChampion.changeToAnimation('idle');
-            });
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                grandPriest.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, grandPriest, 256, 250, 'move')) {
-                    return true;
-                }
-                grandPriest.d = 'down';
-                grandPriest.changeToAnimation('idle');
-            });
-            appendScript(state, `Child, the Spirit King tells me to set you free.`);
-            // the hero stands up and faces the grand priest
-            appendCallback(state, (state: GameState) => {
-                state.hero.action = null;
-                state.hero.d = 'up';
-            });
-            appendScript(state, '{wait:500}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                jadeChampion.speed = 1.5;
-                if (moveNPCToTargetLocation(state, jadeChampion, 256, 300, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'up';
-                jadeChampion.changeToAnimation('idle');
-            });
-            appendScript(state, `I cannot let a Vanara child freely roam the Spirit World.
-                {|}It is my duty as the Jade Champion to protect the Spirit World!`);
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                grandPriest.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, grandPriest, 256, 240, 'move')) {
-                    return true;
-                }
-                grandPriest.d = 'up';
-                grandPriest.changeToAnimation('idle');
-            });
-            appendScript(state, `Silence! My commands are beyond the understanding of mortals.`);
-            appendScript(state, '{wait:500}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                jadeChampion.speed = 1;
-                if (moveNPCToTargetLocation(state, jadeChampion, 208, 276, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'right';
-                jadeChampion.changeToAnimation('idle');
-            });
-            appendScript(state, '{wait:300}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                grandPriest.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, grandPriest, 256, 276, 'move')) {
-                    return true;
-                }
-                grandPriest.d = 'left';
-                grandPriest.changeToAnimation('idle');
-            });
-            appendScript(state, `Champion, seek out the Spirit Beasts.
-                {|}The Spirit Gods shall watch the Vanara child themselves.`);
-            appendScript(state, '{wait:300}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                grandPriest.animationTime += FRAME_LENGTH;
-                if (moveNPCToTargetLocation(state, grandPriest, 256, 250, 'move')) {
-                    return true;
-                }
-                grandPriest.d = 'down';
-                grandPriest.changeToAnimation('idle');
-            });
-            appendScript(state, `Return this Vanara to the gates of our Holy City.
-                {|}This is my command.`);
-            appendScript(state, '{wait:500}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                jadeChampion.speed = 1;
-                if (moveNPCToTargetLocation(state, jadeChampion, 256, 300, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'up';
-                jadeChampion.animationTime += FRAME_LENGTH*2;
-                jadeChampion.changeToAnimation('bow');
+                addBurstEffect(state, jadeChampion, state.hero.area);
+                removeObjectFromArea(state, jadeChampion);
             });
             appendScript(state, '{wait:700}');
-            appendScript(state, `Yes, Grand Priest.`);
-            appendScript(state, '{wait:300}');
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                jadeChampion.speed = 1.25;
-                if (moveNPCToTargetLocation(state, jadeChampion, 220, 332, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'right';
-                jadeChampion.changeToAnimation('idle');
-            });
-            appendScript(state, `Come child, I will take you out of the city.`);
-            // JC moves to cast teleport
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                jadeChampion.speed = 1.25;
-                if (moveNPCToTargetLocation(state, jadeChampion, 240, 332, 'move')) {
-                    return true;
-                }
-                jadeChampion.d = 'right';
-                jadeChampion.changeToAnimation('idle');
-            });
-            npcAndHeroTeleportAnimation(state, jadeChampion);
-            // JC teleports the hero to the Holy City
-            appendCallback(state, (state: GameState) => {
-                enterZoneByTarget(state, 'overworld', 'holyCityCentralGateMarker', null, false, (state) => {
-                    if (state.hero.renderParent && state.hero.renderParent.area !== state.hero.area) {
-                        delete state.hero.renderParent;
-                    }
-                    jadeChampion.x = 90;
-                    jadeChampion.y = 20;
-                    jadeChampion.speed = 1.75;
-                    addObjectToArea(state, state.hero.area, jadeChampion);
-                });
-            });
-            appendScript(state, '{wait:700}');
-            appendCallback(state, (state: GameState) => {
-                state.hero.action = null;
-            });
-            appendScript(state, `Farewell. Please try not to cause any further trouble, little one.`);
-            // JC walks back into the city, then vanishes
-            runPlayerBlockingCallback(state, (state: GameState) => {
-                jadeChampion.animationTime += FRAME_LENGTH;
-                jadeChampion.speed = 1.25;
-                if (moveNPCToTargetLocation(state, jadeChampion, 100, -40, 'move')) {
-                    return true;
-                }
-                jadeChampion.changeToAnimation('idle');
-            });
-            removeObjectFromArea(state, jadeChampion);
-            appendScript(state, '{wait:300}');
-            // remove jadeChampion object
-            // hero.renderParent = jadeChampion -> to make hero vanish during teleports
-            // HIGH: removeobjectfromarea to get rid of champ
-            // or, teleport hero+champ OUTSIDE of gate to south, then have Champ walk north into gate, will auto vanish
-            // HIGH: Make hero+champion disappear any time the burst effect is added.
-            // MED: Reduce delay when teleporting the first time
-            // LOW: Maybe MC+Champ should walk out of temple since it isn't far.
+            // FIX, TODO: Sign to tell player that they have control of their character returned
             return ``;
         },
     },
-    options: [
-        {
-            text: [
-                {
-                    dialogueIndex: undefined,
-                    text: `Speak to the Grand Priest.`,
-                },
-                {
-                    dialogueIndex: undefined,
-                    text: `Something seems terribly wrong, and you are involved somehow.
-                    {|}We will get to the bottom of this.`,
-                },
-            ],
-        },
-    ],
+    options: [],
 };
